requirements.txt
boto3==1.40.69
botocore==1.40.69
celery==5.5.3
Django==5.2.8
djangorestframework==3.16.1
fontTools==4.60.1
Pillow==12.0.0
python-decouple==3.8
Requests==2.32.5
solo==0.0.1
stripe==13.2.0
